Fix typo in "time" variable; add plt.show()
The plt.show() commands are mostly cosmetic, but they suppress the matplotlib return messages you see above the plots. (Each cell in a notebook always displays the return value of the final command in the cell, and for plots, most plotting commands return some sort of object, e.g. `matplotlib.legend.Legend`, but plt.show() doesn't, so the output is a bit cleaner.)